Fixed some warnings.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@21296 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/autostart-prg.c
@@ -289,7 +289,7 @@ int autostart_prg_perform_injection(log_t log)
                 
     /* store data in emu memory */
     for(i = 0; i < prg->size; i++) {
-        mem_store(prg->start_addr + i, prg->data[i]);
+        mem_store((WORD)(prg->start_addr + i), prg->data[i]);
     }
     
     /* now simulate a basic load */
